Simplify loading and move files to interface/
